mission: use mermaidonline for the image
Reduces dependency on this plugin (which is problematic on windows) and also allows rendering in github. Obviously it creates a dependency on mermaidjs online. We could modify to use a hard-coded image and still link to mermaidjs to regenerate. But I think this is good for now. I have kept the original sequence though in comment, so we can regenerate by plugin if needed in future.