Fix payload byte indices in serialization.md
Current span of  payload for MAVLink v2 is described as `10 to (n+10)`, and I believe this is a typo.

For other multi-byte fields like v2 msg_id the byte indices in the left column are inclusive - both byte 7 and byte 9 are part of msg_id. So a payload of one byte should be from byte 10 to byte 10 inclusive. The formula to get one-byte payload then is actually `10 to (n+9)` because n = 1. This might feel a bit strange to programmers used to inclusive-exclusive array boundaries but is correct in this context.

A different explanation is that the byte (n+10) is actually the first byte of the checksum, not the last byte of payload. With current documentation it is unclear which part it belongs to.

This is also true for v1 packet.